Fix vehicle settings ignored when initialization fails
